support react 16
