+++ src/index.js
@@ -1,5 +1,12 @@
 import ElementPortal from './ElementPortal';
+import FiberElementPortal from './FiberElementPortal';
+import LegacyElementPortal from './LegacyElementPortal';
 import withElementPortal from './withElementPortal';
 
-export { withElementPortal };
+export {
+  FiberElementPortal,
+  LegacyElementPortal,
+  withElementPortal
+};
+
 export default ElementPortal;
